refactor(backend): move more stuff out of miscellaneous mod
