+++ apps/backend/src/importer/audiobookshelf.rs
@@ -13,8 +13,10 @@ use serde_json::json;
 use crate::{
     entities::metadata,
     importer::{ImportFailStep, ImportFailedItem, ImportResult},
-    miscellaneous::audiobookshelf_models,
-    models::media::{CommitMediaInput, ImportOrExportMediaItem, ImportOrExportMediaItemSeen},
+    models::{
+        audiobookshelf_models,
+        media::{CommitMediaInput, ImportOrExportMediaItem, ImportOrExportMediaItemSeen},
+    },
     providers::google_books::GoogleBooksService,
     utils::get_base_http_client,
 };
