+++ apps/backend/src/miscellaneous/mod.rs
@@ -1,12 +1,9 @@
 use anyhow::Result;
 use async_graphql::{Enum, InputObject, SimpleObject};
 use enum_meta::{meta, Meta};
-use rust_decimal::Decimal;
 use sea_orm::FromJsonQueryResult;
 use serde::{Deserialize, Serialize};
 use strum::{Display, EnumIter};
-
-use crate::traits::MediaProviderLanguages;
 
 pub mod resolver;
 
@@ -87,84 +84,4 @@ meta! {
             }
         ]
     ), "Items that I want to be reminded about.");
-}
-
-#[derive(Debug, Clone)]
-pub struct CustomService {}
-
-impl MediaProviderLanguages for CustomService {
-    fn supported_languages() -> Vec<String> {
-        ["us"].into_iter().map(String::from).collect()
-    }
-
-    fn default_language() -> String {
-        "us".to_owned()
-    }
-}
-
-pub mod audiobookshelf_models {
-    use super::*;
-
-    #[derive(Debug, Serialize, Deserialize, Clone, Display)]
-    #[serde(rename_all = "snake_case")]
-    pub enum MediaType {
-        Book,
-        Podcast,
-    }
-    #[derive(Debug, Serialize, Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    pub struct ItemProgress {
-        pub progress: Decimal,
-        pub is_finished: bool,
-        pub ebook_progress: Option<Decimal>,
-    }
-    #[derive(Debug, Serialize, Deserialize, Clone)]
-    #[serde(rename_all = "camelCase")]
-    pub struct ItemMetadata {
-        pub title: String,
-        pub id: Option<String>,
-        pub asin: Option<String>,
-        pub isbn: Option<String>,
-        pub itunes_id: Option<String>,
-    }
-    #[derive(Debug, Serialize, Deserialize, Clone)]
-    #[serde(rename_all = "camelCase")]
-    pub struct ItemMedia {
-        pub metadata: ItemMetadata,
-        pub ebook_format: Option<String>,
-        pub episodes: Option<Vec<ItemMetadata>>,
-    }
-    #[derive(Debug, Serialize, Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    pub struct RecentEpisode {
-        pub id: String,
-        pub title: String,
-        pub season: Option<String>,
-        pub episode: Option<String>,
-    }
-    #[derive(Debug, Serialize, Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    pub struct Item {
-        pub id: String,
-        pub name: Option<String>,
-        pub media: Option<ItemMedia>,
-        pub media_type: Option<MediaType>,
-        pub recent_episode: Option<RecentEpisode>,
-        pub user_media_progress: Option<ItemProgress>,
-    }
-    #[derive(Debug, Serialize, Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    pub struct Response {
-        pub library_items: Vec<Item>,
-    }
-
-    #[derive(Debug, Serialize, Deserialize)]
-    pub struct LibrariesListResponse {
-        pub libraries: Vec<Item>,
-    }
-
-    #[derive(Debug, Serialize, Deserialize)]
-    pub struct ListResponse {
-        pub results: Vec<Item>,
-    }
 }
